+++ paper/paper.bib
@@ -1,6 +1,6 @@
 @inproceedings{Seabold2010,
   title = {Statsmodels: Econometric and statistical modeling with python},
-  author = {Skipper Seabold, Josef Perktold},
+  author = {{Seabold}, S. and {Perktold}, J.},
   booktitle = {Proceedings of the 9th Python in Science Conference},
   volume = {57},
   pages = {61},
@@ -9,15 +9,15 @@
 }
 
 @misc{Jones2001,
-  author = {Eric Jones, Travis Oliphant, Pearu Peterson, et al.},
+  author = {{Jones}, E. and {Oliphant}, T. and {Peterson}, P. and et al},
   title = {SciPy: Open source scientific tools for Python},
   year = {2001},
-  url = "{http://www.scipy.org/}"
+  url = {http://www.scipy.org/}
 }
 
-@manual{Pohlert2018,
+@online{Pohlert2018,
   title = {PMCMRplus: Calculate Pairwise Multiple Comparisons of Mean Rank Sums Extended},
-  author = {Thorsten Pohlert},
+  author = {{Pohlert}, T.},
   year = {2018},
   note = {R package version 1.4.1},
   url = {https://CRAN.R-project.org/package=PMCMRplus}
@@ -25,7 +25,7 @@
 
 @inproceedings{McKinney2010,
   title={Data structures for statistical computing in Python},
-  author={Wes McKinney, et al.},
+  author={{McKinney}, W. and et al},
   booktitle={Proceedings of the 9th Python in Science Conference},
   volume={445},
   pages={51-56},
@@ -35,7 +35,7 @@
 
 @book{Oliphant2006,
   title={A guide to NumPy},
-  author={Travis E. Oliphant},
+  author={{Oliphant}, T. E.},
   volume={1},
   year={2006},
   publisher={Trelgol Publishing USA}
