### 2018-12-09 6a268c7
minor fixes
