+++ paper/paper.bib
@@ -9,7 +9,7 @@
 }
 
 @misc{Jones2001,
-  author = {{Jones}, E. and {Oliphant}, T. and {Peterson}, P. and et al},
+  author = {{Jones}, E. and {Oliphant}, T. and {Peterson}, P.},
   title = {SciPy: Open source scientific tools for Python},
   year = {2001},
   url = {http://www.scipy.org/}
@@ -25,7 +25,7 @@
 
 @inproceedings{McKinney2010,
   title={Data structures for statistical computing in Python},
-  author={{McKinney}, W. and et al},
+  author={{McKinney}, W.},
   booktitle={Proceedings of the 9th Python in Science Conference},
   volume={445},
   pages={51-56},
