+++ README.rst
@@ -153,9 +153,3 @@ Credits
 -------
 
 Thorsten Pohlert, PMCMR author and maintainer
-
-Donate
-------
-
-.. image:: https://www.paypalobjects.com/en_US/i/btn/btn_donateCC_LG.gif
-    :target: https://www.paypal.com/cgi-bin/webscr?cmd=_donations&business=K5J3W3WUQ754U&lc=US&currency_code=USD&bn=PP%2dDonationsBF%3abtn_donateCC_LG%2egif%3aNonHosted
