### 2019-01-12 #1120 @katyhuff
Hi @maximtrp thanks for this submission. I'm sorry for the delay. I'm back in action now. Now, @maximtrp , let's ping a few fitting reviewers, with computational statistics and python expertise. We can start the review once two or three of them accept the request to review.

----

Dear potential reviewers (@djmitche @kellieotto @mnarayan @acolum ),

Based on your expertise in computational statistics and python, I would like to ask if you can help us out with a submission to JOSS (The Journal of Open Source Software), where I’m an editor.

Title: Scikit-posthocs
Summary: scikit-posthocs provides statistical post-hoc tests for pairwise multiple comparisons and outliers detection algorithms. https://scikit-posthocs.readthedocs.io/
Topic: Computational Statistics, automated comparison plotting, scikit ecosystem, etc.
Article Proof: https://github.com/openjournals/joss-papers/blob/joss.01120/joss.01120/10.21105.joss.01120.pdf
Submitting author: @maxheld83 (Maksim Terpilowski)
Repository: https://github.com/maximtrp/scikit-posthocs
Version: v0.4.0

@djmitche : Dustin, are you able to and interested in taking on this review?
@kellieotto : Kellie, are you able to and interested in taking on this review?
@mnarayan : Manjari, are you able to and interested in taking on this review?
@acolum : Alyssa, are you able to and interested in taking on this review?

Your prompt response is appreciated. Please see https://joss.readthedocs.io/en/latest/reviewer_guidelines.html for more information about reviewer criteria and the process if it is new to you.
