### 2019-01-12 #1169 @whedon
[REVIEW]: scikit-posthocs: Pairwise multiple comparison tests in Python
**Submitting author:** @maximtrp (<a href="http://orcid.org/0000-0003-2586-4633">Maksim Terpilowski</a>)
**Repository:** <a href="https://github.com/maximtrp/scikit-posthocs" target ="_blank">https://github.com/maximtrp/scikit-posthocs</a>
**Version:** v0.5.1
