### 2019-01-13 maximtrp/scikit-posthocs#14 @djmitche
More information in API documentation
The [API documentation](https://scikit-posthocs.readthedocs.io/en/latest/posthocs_api/) describes the available functions and names their arguments, but does not describe what those arguments mean, what the output is.

As a user of this package, I would be concerned that I had misunderstood what, for example, `g` is in `posthoc_tukey_hsd(x, g[, alpha])`.  I might also struggle to figure out the format of the return value, and thus perhaps mis-interpret the results.  A bit of additional context would give me confidence that I was using the test I thought I was using, and doing so correctly.

I suspect that the arguments are roughly the same for all of the methods of a particular API, so this could be fixed with some preliminary text describing common variables (`a`, `y_col`, etc.) and return values.  A few methods will need additions to the per-method description for unique variables like `g` or `alpha`.  These cases might also benefit from links or citations of specific descriptions of the test (or plot or outlier detection method) that use the same variable names, in case the user's text defines them differently.

(For https://github.com/openjournals/joss-reviews/issues/1169)
