### 2019-01-13 maximtrp/scikit-posthocs#15 @djmitche
Example usage appears only in README
https://scikit-posthocs.readthedocs.io/ does not appear to contain the examples that are featured in the README.  If it's easily done and maintained, it might be helpful for new users to find examples in whichever place they encounter first.

(For https://github.com/openjournals/joss-reviews/issues/1169)
