### 2019-01-13 #1169 @djmitche
In

> Do all archival references that should have a DOI list one (e.g., papers, datasets, software)?

I'm unfamiliar with what "should have a DOI" means.  But I don't see any DOI's in the references.  @mnarayan is that something you're familar with and could look at more deeply?
