### 2019-01-13 maximtrp/scikit-posthocs#16 @djmitche
Tests for plotting, outliers
I see and appreciate a good test suite for the post-hoc tests that make up the "core" of this library, but I don't see any tests for the plotting and outlier APIs.  As a contributor, any change I made to either of those files (perhaps to improve performance, for example) would not be subject to any automated testing and could perhaps introduce errors that would invalidate analyses done with the tools.

I see that the post-hoc tests basically take some input data and test that the output closely matches the expected value, and do so once for each post-hoc test.  That may fail to catch "edge cases" but still successfully executes each function and evaluates the common case.  Doing the same for plotting and outliers should be sufficient.

(For https://github.com/openjournals/joss-reviews/issues/1169)
