### 2019-01-13 maximtrp/scikit-posthocs#17 @djmitche
Add contribution guidelines, bug reporting, support
https://github.com/openjournals/joss-reviews/issues/1169 both suggests having a clear description of how to

 * Contribute to the software
 * Report issues or problems
 * Seek support

Github provides [easy setup](https://github.com/maximtrp/scikit-posthocs/community) for all of this.  Adding some links and text to the README would finish the job.

Note that this does not require you to commit to providing "support' - just make it clear what level of support a user might expect, and where best to go to get that support (github issues, mailing list, email the author, etc.)
