### 2019-01-16 f10dae7
Create CONTRIBUTING.md
