+++ CONTRIBUTING.md
@@ -1,0 +1,15 @@
+# Support 
+
+Currently, all support is provided on GitHub. Please open an issue with your bug, question, or suggestion.
+
+# Bugs
+
+If you have found a bug, open a GitHub issue containing the following information:
+
+* Full error traceback.
+* Steps to reproduce a bug.
+* Dataset you get a bug with.
+
+# Contribution
+
+Your contribution is highly welcome. You may open a pull request or an issue describing an improvement or implementation of new functionality.
