+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,0 +1,28 @@
+---
+name: Bug report
+about: Create a report to help us improve
+title: ''
+labels: bug
+assignees: maximtrp
+
+---
+
+**Describe the bug**
+A clear and concise description of what the bug is.
+
+**To Reproduce**
+Steps to reproduce the behavior:
+1. Go to '...'
+2. Click on '....'
+3. Scroll down to '....'
+4. See error
+
+**Expected behavior**
+A clear and concise description of what you expected to happen.
+
+**System and package information (please complete the following information):**
+ - OS: [e.g. Linux 4.20.0-arch1-1-ARCH x86_64 GNU/Linux]
+ - Package version: [e.g. 0.4.0]
+
+**Additional context**
+Add any other context about the problem here.
