### 2019-01-16 9cd219e
Create CODE_OF_CONDUCT.md
