+++ tests/__init__.py
@@ -1,7 +1,6 @@
 import unittest
 import tests.test_posthocs
 
-
 def posthocs_suite():
     loader = unittest.TestLoader()
     suite = loader.loadTestsFromModule(test_posthocs)
