### 2019-01-18 f78ed88
attempt to fix travis "no display bug"
