### 2019-01-19 maximtrp/scikit-posthocs#15 @maximtrp
I have added examples to the docs. Planning to add examples per each API section.
