### 2019-01-19 maximtrp/scikit-posthocs#17 @maximtrp
Added [CONTRIBUTING.md](https://github.com/maximtrp/scikit-posthocs/blob/master/CONTRIBUTING.md) with some basic information. Also, I have created a **Bug report** template with appropriate sections that should be filled in.
