### 2019-01-19 maximtrp/scikit-posthocs#16 @maximtrp
Added tests for each method of plotting and outliers API.

I have plans to reorganize the tests considering the examples found in scientific papers to ensure the correctness of implementation.
