### 2019-01-19 maximtrp/scikit-posthocs#17 @djmitche
Looks good!
