### 2019-01-19 maximtrp/scikit-posthocs#16 @djmitche
Looks good!  I see that they're nested in the TestPosthocs which might make them a little difficult to find, but that's easy to fix in reorg.  The tests exist, and that's key!
