### 2019-01-21 maximtrp/scikit-posthocs#18 @mnarayan
JOSS Review: Clarify statistical functionality of this package
This relates to requirements stated in https://github.com/openjournals/joss-reviews/issues/1169. Both statement of need and functionality provided by this software needs further clarification.

From reading the README and paper.md alone, I don't immediately understand what scikit-posthoc does compared to statsmodel's multiple testing functionality. It would be helpful to have one or two paragraphs explaining exactly which sorts of multiple testing scenarios an end user should use scikit-posthocs for rather than say just invoking the multiple testing functionality of statsmodels.

The PMCMR package in R has a lot more detail in its documentation. However, it seems they implement a number of use-cases that this package does not yet support.
