+++ docs/source/intro.rst
@@ -1,12 +1,20 @@
 Introduction
 ============
 
-*scikit-posthocs* provides statistical post-hoc tests for pairwise multiple comparisons and outliers detection algorithms.
+`scikit-posthocs` is a Python package which provides post hoc tests for pairwise multiple comparisons that are usually performed in statistical data analysis to assess the differences between group levels if a statistically significant result of ANOVA test has been obtained. `scikit-posthocs` is tightly integrated with Pandas DataFrames and NumPy arrays to ensure fast computations and convenient data import and storage.
+
+Background
+----------
+
+Python statistical ecosystem is comprised of multiple packages. However, it still has numerous gaps and is surpassed by R packages and capabilities. `SciPy <https://www.scipy.org/>`_ (version 1.2.0) offers *Student*, *Wilcoxon*, and *Mann-Whitney* tests which are not adapted to multiple pairwise comparisons. `Statsmodels <http://statsmodels.sourceforge.net/>`_ (version 0.9.0) features *TukeyHSD* test which needs some extra actions to be fluently integrated into a data analysis pipeline. It also has a good helper method `statsmodels.sandbox.stats.multicomp.MultiComparison.allpairtest` which requires an external function (such as `scipy.stats.ttest_ind`) to make multiple pairwise comparisons. It includes `statsmodels.stats.multitest.multipletests` method which makes *p* values corrections for multiple tests. `PMCMRplus <https://rdrr.io/cran/PMCMRplus/>`_ is a very good R package which has no rivals in Python as it offers more than 40 various tests (including post hoc tests) for factorial and block design data. PMCMRplus was an inspiration and a reference for `scikit-posthocs`.
+
+`scikit-posthocs` attempts to improve Python statistical capabilities by offering a lot of parametric and nonparametric post hoc tests along with outliers detection and basic plotting methods.
+
 
 Features
 --------
 
-- Parametric and nonparametric pairwise multiple comparisons tests:
+- Pairwise multiple comparisons parametric and nonparametric tests:
 
   - Conover, Dunn, and Nemenyi tests for use with Kruskal-Wallis test.
   - Conover, Nemenyi, Siegel, and Miller tests for use with Friedman test.
@@ -18,13 +26,13 @@ Features
   - Scheffe test.
   - Tamhane T2 test.
 
-- Plotting (significance plots).
+- Outliers detection tests:
 
-- Outliers detection algorithms:
-
-  - Simple interquartile range (IQR) test.
+  - Simple test based on interquartile range (IQR).
   - Grubbs test.
   - Tietjen-Moore test.
   - Generalized Extreme Studentized Deviate test (ESD test).
 
-  All pairwise tests are capable of p adjustments for multiple pairwise comparisons.
+- Plotting functionality (e.g. significance plots).
+
+All post hoc tests are capable of p adjustments for multiple pairwise comparisons.
