+++ scikit_posthocs/_posthocs.py
@@ -8,32 +8,96 @@ from statsmodels.stats.multicomp import pairwise_tukeyhsd
 from statsmodels.stats.libqsturng import psturng
 from pandas import DataFrame, Categorical, Series
 
-def __convert_to_df(a, val_col, group_col):
+def __convert_to_df(a, val_col = 'val', group_col = 'groups', val_id = None, group_id = None):
+
+    '''
+    Hidden helper method to create a DataFrame with input data for further
+    processing.
+
+    Parameters
+    ----------
+    a : array_like or pandas DataFrame object
+        An array, any object exposing the array interface or a pandas DataFrame.
+        Array must be two-dimensional. Second dimension may vary,
+        i.e. groups may have different lengths.
+
+    val_col : str, optional
+        Name of a DataFrame column that contains numerical values (test
+        variable). Must be specified if `a` is a pandas DataFrame object.
+
+    group_col : str, optional
+        Name of a DataFrame column that contains categorical values (grouping
+        variable). Must be specified if `a` is a pandas DataFrame object.
+
+    val_id : int, optional
+        Index of a column that contains numerical values (test variable).
+        Should be specified if a NumPy ndarray is used as an input. It will be
+        inferred from data, if not specified.
+
+    group_id : int, optional
+        Index of a column that contains categorical values (grouping variable).
+        Should be specified if a NumPy ndarray is used as an input. It will be
+        inferred from data, if not specified.
+
+    Returns
+    -------
+    x : pandas DataFrame
+        DataFrame with input data, `val_col` column contains numerical values and
+        `group_col` column contains categorical values.
+
+    val_col : str
+        Name of a DataFrame column that contains numerical values (test
+        variable). Must be specified if `a` is a pandas DataFrame object.
+
+    group_col : str
+        Name of a DataFrame column that contains categorical values (grouping
+        variable). Must be specified if `a` is a pandas DataFrame object.
+
+    Notes
+    -----
+    Inferrence algorithm for determining `val_id` and `group_id` args is rather
+    simple, therefore it is better to specify them explicitly to prevent errors.
+
+    '''
 
     if isinstance(a, DataFrame):
         x = a.copy()
-        if not all([group_col, val_col]):
-            raise ValueError('group_col, val_col must be explicitly specified')
-    else:
-        x = np.array(a)
+        if not {group_col, val_col}.issubset(a.columns):
+            raise ValueError('Specify correct column names using `group_col` and `val_col` args')
+        return x, val_col, group_col
 
-        if not all([group_col, val_col]):
-            try:
-                groups = np.array([len(a) * [i + 1] for i, a in enumerate(x)])
-                groups = sum(groups.tolist(), [])
-                x = sum(x.tolist(), [])
-                x = np.column_stack([x, groups])
-                val_col = 0
-                group_col = 1
-            except:
-                raise ValueError('array cannot be processed, provide val_col and group_col args')
[…]
