+++ scikit_posthocs/_posthocs.py
@@ -8,10 +8,9 @@ from statsmodels.stats.multicomp import pairwise_tukeyhsd
 from statsmodels.stats.libqsturng import psturng
 from pandas import DataFrame, Categorical, Series
 
-def __convert_to_df(a, val_col = 'val', group_col = 'groups', val_id = None, group_id = None):
+def __convert_to_df(a, val_col=None, group_col=None, val_id=None, group_id=None):
 
-    '''
-    Hidden helper method to create a DataFrame with input data for further
+    '''Hidden helper method to create a DataFrame with input data for further
     processing.
 
     Parameters
@@ -60,6 +59,11 @@ def __convert_to_df(a, val_col = 'val', group_col = 'groups', val_id = None, gro
 
     '''
 
+    if not group_col:
+        group_col = 'groups'
+    if not val_col:
+        val_col = 'vals'
+
     if isinstance(a, DataFrame):
         x = a.copy()
         if not {group_col, val_col}.issubset(a.columns):
@@ -70,6 +74,7 @@ def __convert_to_df(a, val_col = 'val', group_col = 'groups', val_id = None, gro
         grps_len = map(len, a)
         grps = list(it.chain(*[[i+1] * l for i, l in enumerate(grps_len)]))
         vals = list(it.chain(*a))
+
         return DataFrame({val_col: vals, group_col: grps}), val_col, group_col
 
     elif isinstance(a, np.ndarray):
@@ -135,10 +140,9 @@ def __convert_to_block_df(a, y_col, group_col, block_col, melted):
     return x, 'y', 'groups', 'blocks'
 
 
-def posthoc_conover(a, val_col = None, group_col = None, p_adjust = None, sort = True):
+def posthoc_conover(a, val_col=None, group_col=None, p_adjust=None, sort=True):
 
-    '''
-    Post-hoc pairwise test for multiple comparisons of mean rank sums
+    '''Post hoc pairwise test for multiple comparisons of mean rank sums
     (Conover's test). May be used after Kruskal-Wallis one-way analysis of
     variance by ranks to do pairwise comparisons [1]_.
 
@@ -177,8 +181,8 @@ def posthoc_conover(a, val_col = None, group_col = None, p_adjust = None, sort =
 
     Returns
     -------
-    Numpy ndarray or pandas DataFrame (depending on input data type) with p
-    values.
+    result : pandas DataFrame
+        DataFrame with P values.
 
     Notes
     -----
@@ -255,10 +259,9 @@ def posthoc_conover(a, val_col = None, group_col = None, p_adjust = None, sort =
     return DataFrame(vs, index=x_groups_unique, columns=x_groups_unique)
 
 
-def posthoc_dunn(a, val_col = None, group_col = None, p_adjust = None, sort = True):
+def posthoc_dunn(a, val_col=None, group_col=None, p_adjust=None, sort=True):
 
-    '''
-    Post-hoc pairwise test for multiple comparisons of mean rank sums
+    '''Post hoc pairwise test for multiple comparisons of mean rank sums
     (Dunn's test). May be used after Kruskal-Wallis one-way analysis of
     variance by ranks to do pairwise comparisons [1]_, [2]_.
 
@@ -368,10 +371,9 @@ def posthoc_dunn(a, val_col = None, group_col = None, p_adjust = None, sort = Tr
     return DataFrame(vs, index=x_groups_unique, columns=x_groups_unique)
 
 
-def posthoc_nemenyi(a, val_col = None, group_col = None,  dist = 'chi', sort = True):
+def posthoc_nemenyi(a, val_col=None, group_col=None,  dist='chi', sort=True):
 
-    '''
[…]
