### 2019-03-16 8842777
minor fixes
