+++ scikit_posthocs/_posthocs.py
@@ -182,7 +182,7 @@ def posthoc_conover(a, val_col=None, group_col=None, p_adjust=None, sort=True):
     Returns
     -------
     result : pandas DataFrame
-        DataFrame with P values.
+        P values.
 
     Notes
     -----
@@ -198,10 +198,6 @@ def posthoc_conover(a, val_col=None, group_col=None, p_adjust=None, sort=True):
 
     >>> x = [[1,2,3,5,1], [12,31,54, np.nan], [10,12,6,74,11]]
     >>> sp.posthoc_conover(x, p_adjust = 'holm')
-    array([ [-1.        ,  0.00119517,  0.00278329],
-            [ 0.00119517, -1.        ,  0.18672227],
-            [ 0.00278329,  0.18672227, -1.        ]])
-
     '''
 
     def compare_conover(i, j):
@@ -300,8 +296,8 @@ def posthoc_dunn(a, val_col=None, group_col=None, p_adjust=None, sort=True):
 
     Returns
     -------
-    Numpy ndarray or pandas DataFrame of p values depending on input
-    data type.
+    result : pandas DataFrame
+        P values.
 
     Notes
     -----
@@ -318,10 +314,6 @@ def posthoc_dunn(a, val_col=None, group_col=None, p_adjust=None, sort=True):
 
     >>> x = [[1,2,3,5,1], [12,31,54, np.nan], [10,12,6,74,11]]
     >>> sp.posthoc_dunn(x, p_adjust = 'holm')
-    array([[-1.          0.01764845  0.04131415]
-           [ 0.01764845 -1.          0.45319956]
-           [ 0.04131415  0.45319956 -1.        ]])
-
     '''
 
     def compare_dunn(i, j):
@@ -367,7 +359,6 @@ def posthoc_dunn(a, val_col=None, group_col=None, p_adjust=None, sort=True):
 
     vs[tri_lower] = vs.T[tri_lower]
     np.fill_diagonal(vs, -1)
-
     return DataFrame(vs, index=x_groups_unique, columns=x_groups_unique)
 
 
@@ -402,8 +393,8 @@ def posthoc_nemenyi(a, val_col=None, group_col=None,  dist='chi', sort=True):
 
     Returns
     -------
-    Numpy ndarray or pandas DataFrame of p values depending on input
-    data type.
+    result : pandas DataFrame
+        P values.
 
     Notes
     -----
@@ -418,10 +409,6 @@ def posthoc_nemenyi(a, val_col=None, group_col=None,  dist='chi', sort=True):
     --------
     >>> x = [[1,2,3,5,1], [12,31,54, np.nan], [10,12,6,74,11]]
     >>> sp.posthoc_nemenyi(x)
-    array([[-1.        ,  0.02206238,  0.06770864],
-           [ 0.02206238, -1.        ,  0.75361555],
-           [ 0.06770864,  0.75361555, -1.        ]])
-
     '''
 
     def compare_stats_chi(i, j):
@@ -527,7 +514,8 @@ def posthoc_nemenyi_friedman(a, y_col=None, block_col=None, group_col=None, melt
 
     Returns
     -------
-    Pandas DataFrame containing p values.
+    result : pandas DataFrame
+        P values.
 
     Notes
     -----
@@ -655,7 +643,8 @@ def posthoc_conover_friedman(a, y_col=None, block_col=None, group_col=None, melt
 
     Returns
     -------
-    Pandas DataFrame containing p values.
+    result : pandas DataFrame
+        P values.
 
     Notes
     -----
@@ -770,7 +759,8 @@ def posthoc_npm_test(a, val_col=None, group_col=None, sort=False, p_adjust=None)
 
     Returns
     -------
-    Pandas DataFrame containing p values.
[…]
