+++ README.rst
@@ -70,11 +70,17 @@ Package is compatible with Python 2 and Python 3.
 Install
 -------
 
-You can install the package ``pip`` :
-``pip install scikit-posthocs``
+You can install the package using ``pip`` :
+
+.. code:: bash
+
+  pip install scikit-posthocs
+
 
 Examples
 --------
+
+Input data must be 
 
 List or NumPy array
 ~~~~~~~~~~~~~~~~~~~
