### 2019-03-23 maximtrp/scikit-posthocs#18 @mnarayan
@maximtrp how is this going?
