### 2019-03-23 maximtrp/scikit-posthocs#20 @mnarayan
JOSS Review: Clarify Statement of need
This is related to #18 as well. The package introduction and documentation assumes that the audience knows a great deal about pairwise multiple comparisons that arise after conducting an ANOVA. However, these scenarios are quite different from other sorts of multiple testing comparisons that arise.

It would be great to provide a canonical example/scenario that a user can relate to.
Relevant to https://github.com/openjournals/joss-reviews/issues/1169
