### 2019-03-23 #1169 @mnarayan
@maximtrp I can see a lot of code improvements, you are almost there!
My only sticking points are related to clarifying the functionality and description of what this package offers. Even the PMCMRplus package that is the inspiration for yours is a bit opaque. This shouldn't take you too long however.
