+++ scikit_posthocs/_posthocs.py
@@ -21,22 +21,24 @@ def __convert_to_df(a, val_col=None, group_col=None, val_id=None, group_id=None)
         i.e. groups may have different lengths.
 
     val_col : str, optional
-        Name of a DataFrame column that contains numerical values (test
-        variable). Must be specified if `a` is a pandas DataFrame object.
+        Name of a DataFrame column that contains dependent variable values (test
+        or response variable). Values should have a non-nominal scale. Must be
+        specified if `a` is a pandas DataFrame object.
 
     group_col : str, optional
-        Name of a DataFrame column that contains categorical values (grouping
-        variable). Must be specified if `a` is a pandas DataFrame object.
+        Name of a DataFrame column that contains independent variable values
+        (grouping or predictor variable). Values should have a nominal scale
+        (categorical). Must be specified if `a` is a pandas DataFrame object.
 
     val_id : int, optional
-        Index of a column that contains numerical values (test variable).
-        Should be specified if a NumPy ndarray is used as an input. It will be
-        inferred from data, if not specified.
+        Index of a column that contains dependent variable values (test or
+        response variable). Should be specified if a NumPy ndarray is used as an
+        input. It will be inferred from data, if not specified.
 
     group_id : int, optional
-        Index of a column that contains categorical values (grouping variable).
-        Should be specified if a NumPy ndarray is used as an input. It will be
-        inferred from data, if not specified.
+        Index of a column that contains independent variable values (grouping or
+        predictor variable). Should be specified if a NumPy ndarray is used as
+        an input. It will be inferred from data, if not specified.
 
     Returns
     -------
@@ -45,17 +47,17 @@ def __convert_to_df(a, val_col=None, group_col=None, val_id=None, group_id=None)
         `group_col` column contains categorical values.
 
     val_col : str
-        Name of a DataFrame column that contains numerical values (test
-        variable). Must be specified if `a` is a pandas DataFrame object.
+        Name of a DataFrame column that contains dependent variable values (test
+        or response variable).
 
     group_col : str
-        Name of a DataFrame column that contains categorical values (grouping
-        variable). Must be specified if `a` is a pandas DataFrame object.
+        Name of a DataFrame column that contains independent variable values
+        (grouping or predictor variable).
 
     Notes
     -----
     Inferrence algorithm for determining `val_id` and `group_id` args is rather
-    simple, therefore it is better to specify them explicitly to prevent errors.
+    simple, so it is better to specify them explicitly to prevent errors.
 
     '''
 
@@ -154,15 +156,17 @@ def posthoc_conover(a, val_col=None, group_col=None, p_adjust=None, sort=True):
         i.e. groups may have different lengths.
 
     val_col : str, optional
-        Name of a DataFrame column that contains numerical values (test
-        variable). Must be specified if `a` is a pandas DataFrame object.
+        Name of a DataFrame column that contains dependent variable values (test
+        or response variable). Values should have a non-nominal scale. Must be
+        specified if `a` is a pandas DataFrame object.
 
     group_col : str, optional
-        Name of a DataFrame column that contains categorical values (grouping
-        variable). Must be specified if `a` is a pandas DataFrame object.
+        Name of a DataFrame column that contains independent variable values
+        (grouping or predictor variable). Values should have a nominal scale
+        (categorical). Must be specified if `a` is a pandas DataFrame object.
 
     p_adjust : str, optional
-        Method for adjusting p values. See statsmodels.sandbox.stats.multicomp
+        Method for adjusting p values. See `statsmodels.sandbox.stats.multicomp`
[…]
