+++ paper/paper.md
@@ -17,9 +17,11 @@ bibliography: paper.bib
 
 # Summary
 
-``scikit-posthocs`` is a Python package providing multiple pairwise comparison tests (post hocs). Statisticians, data scientists and researchers can use it in a statistical analysis routine to assess the differences between group levels if a statistically significant result of a parametric or nonparametric analysis of variance (ANOVA) test has been obtained.
+``scikit-posthocs`` is a Python package providing multiple pairwise comparison tests (post hocs). Statisticians, data scientists, and researchers will find it useful in a statistical analysis routine to assess the differences between group levels if a statistically significant result of a parametric or nonparametric analysis of variance (ANOVA) test has been obtained.
 
-Most statistical software packages (such as SPSS and Stata) provide a number of post hoc tests. However, Python ecosystem still lacks implementation of these tests for factorial and block design data. Currently, *Student*, *Wilcoxon*, *Mann-Whitney* tests which are not adapted to multiple pairwise comparisons are offered by **SciPy**. **Statsmodels** package includes *TukeyHSD* test which needs extra actions to be fluently integrated into a data analysis pipeline. ``scikit-posthocs`` package is aimed at filling this gap by providing a number of nonparametric and parametric pairwise comparisons tests as well as outliers detection algorithms implemented in Python.
+Most statistical software packages (such as SPSS and Stata) provide a number of post hoc tests. However, Python ecosystem still lacks implementation of these tests for factorial and block design data. Currently, *Student*, *Wilcoxon*, *Mann-Whitney* tests which are not adapted to multiple pairwise comparisons are offered by **SciPy**. **Statsmodels** package includes *TukeyHSD* test which needs extra actions to be fluently integrated into a data analysis pipeline.
+
+``scikit-posthocs`` package is aimed at filling this gap by providing a number of nonparametric and parametric pairwise comparisons tests as well as outliers detection algorithms implemented in Python.
 
 Currently, the following post hoc tests are implemented in this package:
 
