+++ docs/source/index.rst
@@ -1,17 +1,24 @@
-.. scikit-posthocs documentation master file, created by
-   sphinx-quickstart on Fri Oct 12 11:07:23 2018.
-   You can adapt this file completely to your liking, but it should at least
-   contain the root `toctree` directive.
+scikit-posthocs
+===============
 
-Welcome to scikit-posthocs documentation!
-===========================================
+*scikit-posthocs* is a Python package which provides post hoc tests for pairwise multiple comparisons that are usually performed in statistical data analysis to assess the differences between group levels if a statistically significant result of ANOVA test has been obtained.
+
+*scikit-posthocs* is tightly integrated with Pandas DataFrames and NumPy arrays to ensure fast computations and convenient data import and storage.
+
+This package will be useful for statisticians, data analysts, and researchers who use Python in their work.
+
 
 .. toctree::
+   :caption: Documentation
    :maxdepth: 2
 
    Introduction <intro>
    Installation <installation>
-   Examples <examples>
+   Tutorial <tutorial>
+
+.. toctree::
+   :caption: API
+   :maxdepth: 2
 
    Post-hocs API <posthocs_api>
    Plotting API <plotting_api>
