+++ docs/source/tutorial.rst
@@ -1,87 +1,5 @@
-===============
-scikit-posthocs
-===============
-
-.. image:: https://travis-ci.org/maximtrp/scikit-posthocs.svg?branch=master
-    :target: https://travis-ci.org/maximtrp/scikit-posthocs
-.. image:: https://img.shields.io/readthedocs/scikit-posthocs.svg
-    :target: https://scikit-posthocs.readthedocs.io
-.. image:: https://img.shields.io/github/issues/maximtrp/scikit-posthocs.svg
-    :target: https://github.com/maximtrp/scikit-posthocs/issues
-.. image:: https://img.shields.io/pypi/v/scikit-posthocs.svg
-    :target: https://pypi.python.org/pypi/scikit-posthocs/
-.. image:: https://img.shields.io/badge/donate-PayPal-blue.svg
-    :target: https://www.paypal.com/cgi-bin/webscr?cmd=_donations&business=K5J3W3WUQ754U&lc=US&currency_code=USD&bn=PP%2dDonationsBF%3abtn_donateCC_LG%2egif%3aNonHosted
-
-*scikit-posthocs* is a Python package which provides post hoc tests for pairwise multiple comparisons that are usually performed in statistical data analysis to assess the differences between group levels if a statistically significant result of ANOVA test has been obtained.
-
-*scikit-posthocs* is tightly integrated with Pandas DataFrames and NumPy arrays to ensure fast computations and convenient data import and storage.
-
-This package will be useful for statisticians, data analysts, and researchers who use Python in their work.
-
-
-Background
-----------
-
-Python statistical ecosystem is comprised of multiple packages. However, it still has numerous gaps and is surpassed by R packages and capabilities.
-
-`SciPy <https://www.scipy.org/>`_ (version 1.2.0) offers *Student*, *Wilcoxon*, and *Mann-Whitney* tests which are not adapted to multiple pairwise comparisons. `Statsmodels <http://statsmodels.sourceforge.net/>`_ (version 0.9.0) features *TukeyHSD* test which needs some extra actions to be fluently integrated into a data analysis pipeline. `Statsmodels <http://statsmodels.sourceforge.net/>`_ also has good helper methods: ``allpairtest`` (adapts an external function such as ``scipy.stats.ttest_ind`` to multiple pairwise comparisons) and ``multipletests`` (adjusts *p* values to minimize type I and II errors). `PMCMRplus <https://rdrr.io/cran/PMCMRplus/>`_ is a very good R package which has no rivals in Python as it offers more than 40 various tests (including post hoc tests) for factorial and block design data. PMCMRplus was an inspiration and a reference for *scikit-posthocs*.
-
-*scikit-posthocs* attempts to improve Python statistical capabilities by offering a lot of parametric and nonparametric post hoc tests along with outliers detection and basic plotting methods.
-
-
-Features
---------
-
-- Pairwise multiple comparisons parametric and nonparametric tests:
-
-  - Conover, Dunn, and Nemenyi tests for use with Kruskal-Wallis test.
-  - Conover, Nemenyi, Siegel, and Miller tests for use with Friedman test.
-  - Quade, van Waerden, and Durbin tests.
-  - Student, Mann-Whitney, Wilcoxon, and TukeyHSD tests.
-  - Anderson-Darling test.
-  - Mack-Wolfe test.
[…]

[…]
