### 2019-03-30 maximtrp/scikit-posthocs#20 @maximtrp
@mnarayan I have made numerous improvements to code, documentation, and docstrings, as well as drawn a flowchart and written two short tutorials. I hope this is enough.
