### 2019-03-30 maximtrp/scikit-posthocs#18 @maximtrp
@mnarayan Please check out the current state of this repo and its documentation.
