### 2019-03-30 maximtrp/scikit-posthocs#15 @maximtrp
@djmitche I have updated the documentation
