### 2019-03-30 maximtrp/scikit-posthocs#14 @maximtrp
@djmitche API docs are updated. I hope everything is fine now.
