### 2019-03-30 maximtrp/scikit-posthocs#15 @djmitche
:ok_hand: very nice
