### 2019-03-30 maximtrp/scikit-posthocs#14 @djmitche
Looks great!
