### 2019-03-31 maximtrp/scikit-posthocs#18 @mnarayan
The new examples and figures are great!
