### 2019-03-31 maximtrp/scikit-posthocs#20 @mnarayan
The changes look good.
