### 2019-03-31 #1169 @mnarayan
The scikit-posthocs package implements several statistical testing procedures to conduct pairwise multiple comparisons after the global null in an ANOVA has been rejected. It fills a clear gap in the python ecosystem and has been created as an analog to the `PMCMRplus` package in R.

@maximtrp has addressed all my issues and the package implementation meets all the requirements for JOSS.
