+++ paper/paper.md
@@ -17,17 +17,18 @@ bibliography: paper.bib
 
 # Summary
 
-``scikit-posthocs`` is a Python package providing multiple pairwise comparison tests (post hocs). Statisticians, data scientists, and researchers will find it useful in a statistical analysis routine to assess the differences between group levels if a statistically significant result of a parametric or nonparametric analysis of variance (ANOVA) test has been obtained.
+**scikit-posthocs** is a Python package providing multiple pairwise comparison tests (post hocs). Statisticians, data scientists, and researchers will find it useful in a statistical analysis routine to assess the differences between group levels if a statistically significant result of a parametric or nonparametric analysis of variance (ANOVA) test has been obtained.
 
-Most statistical software packages (such as SPSS and Stata) provide a number of post hoc tests. However, Python ecosystem still lacks implementation of these tests for factorial and block design data. Currently, *Student*, *Wilcoxon*, *Mann-Whitney* tests which are not adapted to multiple pairwise comparisons are offered by **SciPy**. **Statsmodels** package includes *TukeyHSD* test which needs extra actions to be fluently integrated into a data analysis pipeline.
+Most statistical software packages (such as SPSS and Stata) provide a number of post hoc tests. However, Python ecosystem still lacks implementation of these tests for factorial and block design data. Currently, *Student*, *Wilcoxon*, *Mann-Whitney* tests which are not adapted to multiple pairwise comparisons are offered by *SciPy*. *Statsmodels* package includes *TukeyHSD* test which needs extra actions to be fluently integrated into a data analysis pipeline.
 
 ``scikit-posthocs`` package is aimed at filling this gap by providing a number of nonparametric and parametric pairwise comparisons tests as well as outliers detection algorithms implemented in Python.
 
-Currently, the following post hoc tests are implemented in this package:
+Currently, the following tests are implemented in this package:
 
-1. Conover, Dunn, and Nemenyi tests (for use with Kruskal-Wallis test).
-2. Conover, Nemenyi, Siegel, and Miller tests (for use with Friedman test).
-3. Quade, van Waerden, Durbin, Student, Mann-Whitney, Wilcoxon, TukeyHSD, Anderson-Darling, Mack-Wolfe, Nashimoto and Wright (NPM), Scheffe, and Tamhane T2 tests.
+1. *Parametric* pairwise multiple comparisons tests: Scheffe, Student T, Tamhane T2, and TukeyHSD test.
+2. *Non-parametric* tests for factorial design: Conover, Dunn, Dwass-Steel-Critchlow-Fligner, Mann-Whitney, Nashimoto-Wright (NPM), Nemenyi, van Waerden, and Wilcoxon test.
+3. *Non-parametric* tests for block design: Conover, Durbin and Conover, Miller, Nemenyi, Quade, Siegel test.
+4. Additional tests: Anderson-Darling, Mack-Wolfe, Hayter (OSRT) test.
 
 ``scikit-posthocs`` provides tests for outliers detection: interquartile range (IQR) test, Grubbs test, Tietjen-Moore test, and generalized (extreme Studentized deviate) ESD test. It also has plotting functionality to present the results of pairwise comparisons as a heatmap (significance plot, see figure).
 
