### 2019-03-31 5e5ac28
conover-friedman test upd
