+++ tests/test_posthocs.py
@@ -135,13 +135,13 @@ class TestPosthocs(unittest.TestCase):
     def test_posthoc_conover_friedman(self):
 
         results = sp.posthoc_conover_friedman(self.df_bn)
-        p_results = np.array([[-1.000000, 0.935333, 0.268619, 0.339721, 0.339721, 0.060540, 0.628079],
-                              [0.935333, -1.000000, 0.302605, 0.380025, 0.380025, 0.070050, 0.685981],
-                              [0.268619, 0.302605, -1.000000, 0.871144, 0.871144, 0.380025, 0.519961],
-                              [0.339721, 0.380025, 0.871144, -1.000000, 1.000000, 0.302605, 0.628079],
-                              [0.339721, 0.380025, 0.871144, 1.000000, -1.000000, 0.302605, 0.628079],
-                              [0.060540, 0.070050, 0.380025, 0.302605, 0.302605, -1.000000, 0.141412],
-                              [0.628079, 0.685981, 0.519961, 0.628079, 0.628079, 0.141412, -1.000000]])
+        p_results = np.array([[-1.000000, 0.9325836, 0.2497915, 0.3203414, 0.3203414, 0.0517417, 0.6136576],
+                              [0.9325836, -1.000000, 0.28339156, 0.36072942, 0.36072942, 0.06033626, 0.67344846],
+                              [0.2497915, 0.28339156, -1.000000, 0.8657092, 0.8657092, 0.3607294, 0.5026565],
+                              [0.3203414, 0.36072942, 0.8657092, -1.000000, 1.000000, 0.2833916, 0.6136576],
+                              [0.3203414, 0.36072942, 0.8657092, 1.000000, -1.000000, 0.2833916, 0.6136576],
+                              [0.0517417, 0.06033626, 0.3607294, 0.2833916, 0.2833916, -1.000000, 0.1266542],
+                              [0.6136576, 0.67344846, 0.5026565, 0.6136576, 0.6136576, 0.1266542, -1.000000]])
         self.assertTrue(np.allclose(results, p_results))
 
     def test_posthoc_miller_friedman(self):
