+++ scikit_posthocs/_posthocs.py
@@ -2485,98 +2485,3 @@ def posthoc_dscf(a, val_col=None, group_col=None, sort=False):
 
     np.fill_diagonal(vs, -1)
     return DataFrame(vs, index=groups, columns=groups)
-
-
-def osr_test(a, val_col=None, group_col=None, sort=False):
-
-    '''One-Sided Studentised Range Test. Performs Hayter's one-sided studentised
-    range test against an ordered alternative for normal data with equal
-    variances [1]_.
-
-    Parameters
-    ----------
-    a : array_like or pandas DataFrame object
-        An array, any object exposing the array interface or a pandas
-        DataFrame.
-
-    val_col : str, optional
-        Name of a DataFrame column that contains dependent variable values (test
-        or response variable). Values should have a non-nominal scale. Must be
-        specified if `a` is a pandas DataFrame object.
-
-    group_col : str, optional
-        Name of a DataFrame column that contains independent variable values
-        (grouping or predictor variable). Values should have a nominal scale
-        (categorical). Must be specified if `a` is a pandas DataFrame object.
-
-    sort : bool, optional
-        If True, sort data by block and group columns.
-
-    Returns
-    -------
-    P value, statistic.
-
-    Notes
-    -----
-    P values are computed from the Tukey distribution.
-
-    References
-    ----------
-    .. [1] Hayter, A.J.(1990) A One-Sided Studentised Range Test for Testing
-        Against a Simple Ordered Alternative, Journal of the American
-        Statistical Association, 85, 778-785.
-
-    Examples
-    --------
-    >>> import scikit_posthocs as sp
-    >>> import pandas as pd
-    >>> x = pd.DataFrame({"a": [1,2,3,5,1], "b": [12,31,54,62,12], "c": [10,12,6,74,11]})
-    >>> x = x.melt(var_name='groups', value_name='values')
-    >>> sp.osr_test(x, val_col='values', group_col='groups')
-
-    '''
-
-    x, _val_col, _group_col = __convert_to_df(a, val_col, group_col)
-
-    if not sort:
-        x[_group_col] = Categorical(x[_group_col], categories=x[_group_col].unique(), ordered=True)
-
-    x.sort_values(by=[_group_col], ascending=True, inplace=True)
-    groups = np.unique(x[_group_col])
-    x_grouped = x.groupby(_group_col)[_val_col]
-
-    xi = x_grouped.mean()
-    ni = x_grouped.count()
-    k = x_groups_unique.size
-    n = len(x.index)
-    df = n - k
-
-    sigma2 = 0
-    c = 0
-
-    for i in range(k):
-        for j in range(ni[i]):
-            c += 1
-            sigma2 += (x[_val_col].iat[c] - xi[i])**2 /df
-
-    sigma = np.sqrt(sigma2)
-
-    def compare(i, j):
-        dif = xi.loc[groups[j]] - xi.loc[groups[i]]
-        A = sigma / np.sqrt(2) * np.sqrt(1 / ni[groups[j]] + 1 / ni[groups[i]])
-        qval = np.abs(dif) / A
-        return p
-
-    vs = np.zeros((groups.size, groups.size), dtype=np.float)
-    tri_upper = np.triu_indices(vs.shape[0], 1)
-    tri_lower = np.tril_indices(vs.shape[0], -1)
-    vs[:,:] = 0
-
-    combs = it.combinations(range(groups.size), 2)
-
-    for i,j in combs:
-        vs[i, j] = compare(groups[i], groups[j])
-
-    stat = np.max(vs)
-    pval = psturng(stat, k, df)
-    return pval, stat
