### 2019-04-14 #1169 @katyhuff
Thanks so much for the ping @djmitche for the ping! It's a hectic end of the semester and I missed the notification that the comments had been handled.

Thanks @mnarayan and @djmitche for your excellent reviews. The care you took and attention to detail that you supplied is what makes JOSS possible.
