### 2019-04-14 #1169 @katyhuff
Thank you @maximtrp for a strong submission and for engaging actively in the review process! I have looked over the paper and have conducted a high level review of the code itself. Everything looks ship-shape to me and we are nearly ready to move forward with acceptance. That said, if you have a moment, I would like to see all existing references contain a DOI where possible. Some of the citations you've chosen are a bit older, and so they don't have DOIs. This is reasonable, though Pandas and SciPy both have options for newer citations if you are so inclined to use those instead. This is optional, in my opinion, since you seem to have used the preferred citations for all packages (which happen to not contain DOIs).

At this point, it'll be good for you as well to to double check the paper, review any lingering details in your code/readme/etc., and then make an archive of the reviewed software in Zenodo/figshare/other service. Please be sure that your created DOI metadata (title, authors, etc.) matches this JOSS submission. Once that's complete, please update this thread with the DOI of the archive, and I'll move forward with accepting the submission! Until then, now is your moment for final touchups!
