### 2019-04-14 #1169 @maximtrp
@katyhuff I would like to keep these references as they are indeed the preferred ones. I have created a release of *scikit-posthocs* on Zenodo. Here is the DOI: [10.5281/zenodo.2639732](https://zenodo.org/record/2639732). Everything else is okay. Thank you!
