### 2019-04-14 #1169 @katyhuff
Thanks @maximtrp for your submission, patience, and cooperation with the process!

Thank you again @mnarayan and @djmitche for your detailed reviews!

This paper should be ready to accept @openjournals/joss-eics - over to you! (Please note the discussion about DOIs, above.)
