### 2019-04-14 #1169 @labarba
@maximtrp — I noticed the version on your Zenodo archive didn't match what we had here, went to your repo, and see your version tag as v0.5.1 ... so I updated the version. Please confirm that is what we want.
