### 2019-04-14 #1169 @maximtrp
@labarba That's right. It should be v0.5.1. Thank you!
