### 2019-04-14 #1169 @labarba
Big thanks to our editor: @katyhuff, and the reviewers: @mnarayan, @djmitche — your contribution to JOSS is invaluable! 🙏
